ENH: Data specified at command line are loaded into data dialog.
Looks at all unparsed arguments, and determines first which of
them are file paths.

If there is only one argument, and that argument is a MRML scene,
the load-data-dialog is bypassed and the scene is loaded directly.

These arguments can be individual files, or entire directories.
If the path is recognized as a directory, its contents are
traversed recursively and added to the data dialog.

See http://na-mic.org/Mantis/view.php?id=2363

Part of #2363