BUG: Fixes incorrect loading of some dependencies in extensions.
Loadable dependencies located in the base Slicer are now found there,
and only dependencies corresponding to the extension will
have the @Slicer_BUNDLE_EXTENSIONS_LOCATION@ prefix prepended.

This fix is somewhat of a hack - this will be amended pending the
resolution of issues #2390 and #2391.  Once the slicer build
directory is no longer hard-coded it will be easier to replace the
paths to dependencies.[1] Also, python code bundled with an extension
is still inaccessible to the python search path on Macs, which should be
fixed during the packaging process.[2]
[1] http://na-mic.org/Mantis/view.php?id=2390
[2] http://na-mic.org/Mantis/view.php?id=2391

See http://na-mic.org/Bug/view.php?id=2358

Partially fixes #2358