WIP: Adding scaling to transforms module
Currently, this adds an extra set of sliders to the Transforms module.
An arbitrary factor (f = 1.1) is used to perform scaling.  If a slider is at
position 0, no scale is made in that direction - a slider value of 2 would
scale by f^2, and a slider value of -2 would scale by f^-2.  This works well
with the current qMRMLLinearTransformSlider because the default rests on 0,
allowing slider values from -MAX to MAX to make sense.

TODO:  The value of the slider should reflect the scale at that time.
The value of 1 should be the default slider position.  The minimum value
should be the reciprocal of the maximum value, with a checkbox allowing the user
to invert the model over the corresponding axis. Therefore, the domain of the
sliders should be (0,1] at a logarithmic scale and [1, \inf) at a linear scale.